* Use SqlToolsService built on .NET Core 2.0 and a build script updates (#131)

* Bump version to 1.0.0a19

* Use .NET Core 2.0 RTM built sqltoolsservice

* Add build script to upload to azure blob storage

* Upgrade to VS 2017

* Remove 3.3 as supported Python version

* Fix perf issue where main event loop takes 100% of CPU (#132)

Fix perf issue where main event loop takes 100% of CPU

We have a 2 threads:
Thread #1 runs in a loop polling the response queue
Thread #2 runs in a loop decoding responses from the sqltoolsservice over stdout and posting them to the response queue

Since thread #1 doesn't sleep, it's takes 100% CPU. In addition, running python 2.7 on windows, #2 doesn’t preempt the CPU due to #1 taking all of the CPU cycles, so no response is processed.

Fix is simple – thread #1 needs to sleep so thread #2 can get scheduled and get it’s work done.

* Refine event loop perf fix in main.py

Refine event loop perf fix in main.py

* Fixing regular expression

Previous regex would result in release:a1 and release_version: 12.
Modified the regex for part Release to only pick up lower case letters.

* Adding missing forward slash on test pypi url

* fixing typos/grammar (#138)

fixing typos/grammar.

* Updating to release version 1.0.0a20.

* Create doc for official msft docs page

* Updated documentation page with usage_guide

* Added link to download adventureworks

* Updated with sqlcmd usage

* Added in run and cloud shell support

* Update/consolidate linux install (#153)

* universal linux wheel gen and setup update.

* Updating version cfg.

* Updating sqltoolsservice container.

* Updating spacing for flake8.

* Updating team email. (#154)

* Fixing resource warning for sqltoolsservice (#158)

* Fixing resource warning for sqltoolsservice by closing stdout after killing process.

* Shortening sleep time during shutdown.

* Fixing missing bracket.

* Updating doc's after repro rename (#160)

* Updating files after repro rename.

* Fixing flake8 issues.

* Ron/platform wheels support (#161)

* Merge release 1.0.0a21 (#155)

* Use SqlToolsService built on .NET Core 2.0 and a build script updates (#131)

* Bump version to 1.0.0a19

* Use .NET Core 2.0 RTM built sqltoolsservice

* Add build script to upload to azure blob storage

* Upgrade to VS 2017

* Remove 3.3 as supported Python version

* Fix perf issue where main event loop takes 100% of CPU (#132)

Fix perf issue where main event loop takes 100% of CPU

We have a 2 threads:
Thread #1 runs in a loop polling the response queue
Thread #2 runs in a loop decoding responses from the sqltoolsservice over stdout and posting them to the response queue

Since thread #1 doesn't sleep, it's takes 100% CPU. In addition, running python 2.7 on windows, #2 doesn’t preempt the CPU due to #1 taking all of the CPU cycles, so no response is processed.

Fix is simple – thread #1 needs to sleep so thread #2 can get scheduled and get it’s work done.

* Refine event loop perf fix in main.py

Refine event loop perf fix in main.py

* Fixing regular expression

Previous regex would result in release:a1 and release_version: 12.
Modified the regex for part Release to only pick up lower case letters.

* Adding missing forward slash on test pypi url

* fixing typos/grammar (#138)

fixing typos/grammar.

* Updating to release version 1.0.0a20.

* Create doc for official msft docs page

* Updated documentation page with usage_guide

* Added link to download adventureworks

* Updated with sqlcmd usage

* Added in run and cloud shell support

* Update/consolidate linux install (#153)

* universal linux wheel gen and setup update.

* Updating version cfg.

* Updating sqltoolsservice container.

* Updating spacing for flake8.

* Updating team email. (#154)

* Updating mssqltoolsservice to be integrated as a package of mssqlscripter.

* Updating sqltoolsservice to be loaded from the repro instead of storage account.

* Fix index file generation for daily storage account.

* Fixing manylinux1 tag.

* Updating platform tag for win x64.

* Renaming sqltoolsservice win x64 folder.

* Adding platform tags for win_amd64, manylinux1_x86_64, manylinux1_i686.

* version bumping to 1.0.0a22.

* Flake8 format fixes.

* Erroring out when build receives invalid flag.

* Fixing tag for win64

* Update libunwind8 install for CentOS

* Ron/sqltoolsservice update (#163)

* Updating sqltoolsservice with self contained version.

* Refreshing sqltoolsservice again.

* Making mssql-scripter executable and adding null checks in main.py

* Adding clean up step to remove build directory after each build.

* Fixing path for build directory.

* Removing 'pypi' from upload step.

* Flake 8 extra line fix.

+++ mssqlscripter/main.py
@@ -18,7 +18,7 @@ import mssqlscripter.scripterlogging as scripterlogging
 import mssqlscripter.argparser as parser
 import mssqlscripter.scriptercallbacks as scriptercallbacks
 import mssqlscripter.sqltoolsclient as sqltoolsclient
-import mssqltoolsservice
+import mssqlscripter.mssqltoolsservice as mssqltoolsservice
 
 logger = logging.getLogger(u'mssqlscripter.main')
 
@@ -101,15 +101,21 @@ def main(args):
                     sys.stdout.write(line)
 
     finally:
+        if sql_tools_client:
+            sql_tools_client.shutdown()
 
-        sql_tools_client.shutdown()
-        tools_service_process.kill()
-        # 1 second time out, allow tools service process to be killed.
-        time.sleep(1)
-        # None value indicates process has not terminated.
-        if not tools_service_process.poll():
-            sys.stderr.write(
-                u'Sql Tools Service process was not shut down properly.')
+        if tools_service_process:
+            tools_service_process.kill()
+            # 1 second time out, allow tools service process to be killed.
+            time.sleep(.1)
+            # Close the stdout file handle or else we would get a resource warning (found via pytest).
+            # This must be closed after the process is killed, otherwise we would block because the process is using
+            # it's stdout.
+            tools_service_process.stdout.close()
+            # None value indicates process has not terminated.
+            if not tools_service_process.poll():
+                sys.stderr.write(
+                    u'Sql Tools Service process was not shut down properly.')
         try:
             # Remove the temp file if we generated one.
             if temp_file_path:
